Implement download to temporary file
Instead of directly downloading into the target file, download to a
temporary file and move it to its final location/name when the download
is complete. This avoids leaving a partially downloaded file in its
final requested location and name.

Also provides a wait for waiting for the client to complete its
operation after stopping is requested.

Bump to version 1.0.4.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>