Implemented seeder mode and interruptible initialization
Implemented an initial seeder mode for users of the Torrent classes that
disables validation. If the torrent is created and automatically used,
or if someone wants to manually disable validation of the local data,
this mode short-circuits the piece hashing and validation during the
initialization of the torrent.

This initialization has also been made interruptible by being detached
in a background thread while the client can be interrupted with
Client.stop(). During initialization, the client is in the new
VALIDATING ClientState.

Also adjusted logging levels and default client parameters for unchoking
frequency and voluntary outbound connections to improve client
performance on start up.

Signed-off-by: David Giffin <david@etsy.com>
Acked-by: Maxime Petazzoni <mpetazzoni@turn.com>