Add support for multi-file torrent storage
This is the first step in supporting multi-file torrents. It provides
support for a multi-file based torrent data byte storage.

The abstracted byte storage has been extracted into a TorrentByteStorage
interface for offset+length based access to the binary file data. The
existing implementation was renamed FileStorage and implements this
interface, and a new FileCollectionStorage implements the same interface
but relying on a collection of FileStorage and intelligently
orchestrates the read/write operations on the impacted ones as required.

The original code was contributed by David Giffin <david@etsy.com> and
heavily reworked for cleaner and easier to understand code, which
remains one of the objectives of the ttorrent library's codebase.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>