Improve torrent creator and fix path expansion
The main() function of the Torrent class now allows to both read
information from a torrent file and create a new torrent file, for
single and multi-file torrents.

This change also fixes a bug in the parent path file expansion and
ensures the underlying storage files point to the correct locations in
all circumstances.

Code for multi-file hashing and multi-file torrent creation was
contributed by David Giffin <david@etsy.com>.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>