Change AnnounceResponseListener interface
Modified the AnnounceResponseListener interface to get split parameters
for leechers, seeders, interval and peer list instead of the BEValue
map. This is in preparation of the upcoming implementation of UDP
tracker support, which uses a different, binary-based announce protocol.

This also allows for a great deal of simplification in the Client's code
when parsing peers, as this logic is moved in a simpler form into the
Announce class.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>