Start refactoring of announcers
In preparation for the implementation of UDP tracker support through a
different implementation of the Announce class, proceed to some code
refoctoring of the Announce class hierarchy.

Announce related stuff is now in its own subpackage
com.turn.ttorrent.client.announce. Announce is now a base abstract class
for announcers, and HTTPAnnounce implements the announcer to talk with
HTTP trackers.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>