B-encoding tool suite improvements
Add a helper function to the BDecoder to decode directly from a
ByteBuffer instead of an InputStream (uses ByteArrayInputStream and an
AutoCloseInputStream to get an input stream from the ByteBuffer's
backing byte array).

Add a helper function to the BEncoder to encode directly into a
ByteBuffer instead of into an OutputStream (uses ByteArrayOutputStream
and ByteBuffer.wrap on the backing array of the output stream).

Add getShort() from a BEValue, returning the shortValue() from the
decoded Number.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>