Peer class improvements
Add a Peer constructor that only takes an IP address and a port number
to create a peer-ID-less peer.

Add a helper function that returns a formatted, short peer ID in the
form "..LAST_6_CHARS_OF_PEER_ID" that is used in various peer ID
outputs.

Add a getRawIp() method that returns the peer's IP address as a byte
array in network byte order.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>