Handle outbound peer connections with a thread pool executor
Use a thread pool executor to connect to remote peers instead of taping
the announce thread to initiate outbound connections (since that's where
the peer discovery event comes from). This allows for the announce
thread to be gracefully stopped without delay with the clients request
graceful exit, which in turns allow for the proper sending of a STOPPED
announce event to the tracker when the client exits.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>