Use jargs command-line argument parsing
Add dependency to JArgs for command-line argument parsing and use it in
the Client and Tracker classes (Torrent utility to follow). It is also
now possible to specify which interface the client will bind itself too
instead of the default interface matching the HOSTNAME environment
variable (fixes issue #11). If not specified, it will bind to localhost
by default.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>