Fix parsing of the numwant announce request URL parameter (fixes #19)
The numwant parameter was not correctly considered as one of the numeric
announce request URL parameters and thus not converted to a numeric
value in the map of message parameters in TrackerService, which lead to
a downstream parsing error in HTTPAnnounceRequestMessage that resulted
in HTTP 400 Bad Request errors with clients that passed a numwant
parameter.

Thanks to Carlos Prados and Anatoli Vladev for tracking this bug down.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>