Fixed RateComparator to never return 0
The rate comparator could return 0, but since Rate doesn't implement
equals() it would lead to strange behavior when building ordered sets
using a RateComparator. This could lead to the number of bound peers
available for unchoked to be incorrect which could prevent the start of
the download when the swarm gets large.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>