Switch to NIO non-blocking sockets
Using blocking sockets meant that to stop a PeerExchange one had to
forcefully interrupt the incoming and outgoing communication threads.
Unfortunately if that interruption came while the incoming thread was
reading or writing on the underlying FileStorage's FileChannel, it would
get immediately closed, forcing the client to stop while only one peer
had disconnected.

This rather large changeset introduces the switch to NIO, non-blocking
sockets for the peer exchanges so that the two threads of each peer
exchange don't need to be forcefully interrupted. The socket channel is
simply closed and their reads fail gracefully at a controlled time.

It also allows to do pure zero-copy interaction with these sockets
through the use of direct byte buffers.

This change also contains a fix for the allocated size of BitTorrent
keep-alive messages, prevents seeders for initiating outbound
connections (it's not their role) and fixes a potential dead-lock on the
peer exchange lock in some forceful peer exchange unbinding situations.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>