Attempt to read the error message returned by the tracker (fixes #26)
When the tracker returns an error code, we were not trying to read the
tracker error message (B-encoded) from the error stream, if it was
available.

This change makes sure we get the correct error message from the tracker
in error scenarios, and makes sure all streams are correctly closed when
we do so to avoid resource leaks.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>