Don't attempt to stop the full client on peers' IOExceptions
Recoverable IOExceptions from single peers actually can happen if a
message is being sent to a peer at the same time it disconnects. The
previous way of handling this was to kill the client in the belief that
the IOException resulted from a local, unrecoverable error, and could
actually result in a dead lock trying to stop the client.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>