Fix HAVE announce bug
When a peer completes the download of a piece, it sends a HAVE message
to all connected peers to inform them that it now has this piece.
Because the same created message was used to send to all peers, all
sending threads were accessing the same ByteBuffer when sending and
some threads would see the buffer has containing no data because the
buffer's position was changed by other threads.

This fixes the problem by always returning a new ByteBuffer view from
the message's underlying data buffer. They are backed by the same data,
but the buffer's position, mark and limit are independent.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>