Implement a basic end-game scenario
Implement a end-game by making the best use of all ready peers even if
pieces have already been requested from other peers. Right now this is
not limited to the end of the download and can happen at anytime, trying
to make the most of the network bandwidth.

The first peer that gets the full piece back wins, the rest of the
blocks will be discarded.

This change required a bit more synchronization plumbing around pieces
and ongoing requests.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>