Dyanimcally compute requestedPieces instead of storing it
We were storing a BitSet of the pieces that we had requested on the
SwarmHandler, but this was causing problems because we weren't clearing
the bits on errors and we quickly re-request pieces then.  The net
effect was that a download would very quickly get to 80% complete, then
begin slowing down, and I never saw one get past 95% to even get into
end-game mode.

This patch makes us compute requestedPieces on demand, which is a bit
inefficient in terms of computation, but serves as a much better source
of truth for what pieces have already been requested and overall makes
downloads complete much faster.