Allow explicit start of peer message exchange
This change fixes a race condition in peer message exchange. First
incoming messages can be lost because incoming thread is started before
it is possible to register a message listener. There is a high
probability of missing the peer’s bitfield message immediately after
handshake and as a consequence downloads will not resume.

This fix is required to resolve issue #91
https://github.com/mpetazzoni/ttorrent/issues/91